+++ src/main/kotlin/org/example/onlinestoreapi/strategies/DiscountStrategy.kt
@@ -3,5 +3,5 @@ package org.example.onlinestoreapi.strategies
 import org.example.onlinestoreapi.entities.Product
 
 fun interface DiscountStrategy {
-    fun applyDiscount(product: Product, discount: Long): Product
+    fun applyDiscount(product: Product, discount: Int): Product
 }
